fix warnings
Former-commit-id: 3eee7ce14b42eb5d0660e558ea96f4223e525357 [formerly eb62b3c050c0a7560bb0b55660ec6f9ef8a4ce10] [formerly 8310c2734abb6a61eae9bcacdfdecd8c065c4cc3 [formerly 94954e03de570cd82ea188ef21f0ad768273fe1b]]
Former-commit-id: 8cef5f4a0a1d5a8aa37a4ad758c3b69b1ab238ed [formerly 9e165bf213c8f88a93385848ebed6188927f4e2c]
Former-commit-id: 98fca552e5c9886c246bb905e05dc780c058ee32

+++ Frame/SDK/Core/AFDataTable.cpp
@@ -100,7 +100,7 @@ namespace ark
         return true;
     }
 
-    int AFDataTable::GetColType(int col) const
+    int AFDataTable::GetColType(size_t col) const
     {
         assert(col < mxColTypes.size());
         return mxColTypes[col];
